feat: Context y view #168
