+++ src/apps/TeacherApp.tsx
@@ -19,6 +19,7 @@ import { NoLudicaProvider } from "../activity/contexts/noLudicaContext/NoLudicaP
 import { BenefitAPIProvider } from "../teacher/contexts/benefitsAPIContext/BenefitAPIProvider";
 import { BenefitUIProvider } from "../teacher/contexts/benefitsUIContext/BenefitUIProvider";
 import { ArbolDecisionProvider } from "../activity/contexts/arbolDecisionContext/ArbolDecisionProvider";
+import { DesafioClasificacionGameProvider } from "../shared/contexts/gamesContext/desafioClasificaionContext/DesafioClasificaionGameProvider";
 //VIEWS
 import ActivitiesView from "../teacher/views/ActivitiesView/ActivitiesView";
 import BenefitsListView from "../teacher/views/benefitsView/benefitsViewList/BenefitsListView";
@@ -52,25 +53,27 @@ const TeacherApp = () => {
                             <PreguntadosProvider>
                               <OrdenarSecuenciaProvider>
                                 <DesafioClasificacionProvider>
-                                  <MemoramaProvider>
-                                    <NoLudicaProvider>
-                                      <ArbolDecisionProvider>
-                                        <BenefitAPIProvider>
-                                          <BenefitUIProvider>
-                                            <motion.div
-                                              key="dashboardTeacher"
-                                              initial={{ opacity: 0 }}
-                                              animate={{ opacity: 1 }}
-                                              exit={{ opacity: 0 }}
-                                              transition={{ duration: 0.3 }}
-                                            >
-                                              <TeacherDashboardPage />
-                                            </motion.div>
-                                          </BenefitUIProvider>
-                                        </BenefitAPIProvider>
-                                      </ArbolDecisionProvider>
-                                    </NoLudicaProvider>
-                                  </MemoramaProvider>
+                                  <DesafioClasificacionGameProvider>
+                                    <MemoramaProvider>
+                                      <NoLudicaProvider>
+                                        <ArbolDecisionProvider>
+                                          <BenefitAPIProvider>
+                                            <BenefitUIProvider>
+                                              <motion.div
+                                                key="dashboardTeacher"
+                                                initial={{ opacity: 0 }}
+                                                animate={{ opacity: 1 }}
+                                                exit={{ opacity: 0 }}
+                                                transition={{ duration: 0.3 }}
+                                              >
+                                                <TeacherDashboardPage />
+                                              </motion.div>
+                                            </BenefitUIProvider>
+                                          </BenefitAPIProvider>
+                                        </ArbolDecisionProvider>
+                                      </NoLudicaProvider>
+                                    </MemoramaProvider>
+                                  </DesafioClasificacionGameProvider>
                                 </DesafioClasificacionProvider>
                               </OrdenarSecuenciaProvider>
                             </PreguntadosProvider>
