+++ src/shared/components/Games/DesafioClasificacion/DesafioClasificacionGame.module.css
@@ -1,0 +1,429 @@
+@import "../../../../shared/styles/colors.css";
+.decisionsSection {
+  background: white;
+  border-radius: 12px;
+  padding: 1.5rem;
+  margin-bottom: 2rem;
+  border: 1px solid #e5e7eb;
+}
+
+.decisionsSectionHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1.5rem;
+}
+
+.decisionsSectionTitle {
+  font-size: 1.1rem;
+  font-weight: 600;
+  color: var(--color-text);
+  margin: 0;
+}
+
+.decisionsInfo {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  color: #6b7280;
+  font-size: 0.9rem;
+}
+.conceptsPool {
+  background: #f7fafc;
+  border-radius: 12px;
+  padding: 1rem;
+  border: 2px solid #e2e8f0;
+  min-height: 80px;
+  transition: all 0.2s ease;
+}
+
+.conceptsPool:hover {
+  border-color: var(--color-stat-4);
+}
+
+.conceptsList {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.75rem;
+  justify-content: center;
+  min-height: 60px;
+  align-items: center;
+}
+
+.conceptChip {
+  padding: 0.5rem 1rem;
+  background: white;
+  border: 2px solid #cbd5e0;
+  border-radius: 20px;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: var(--color-text);
+  cursor: grab;
+  user-select: none;
+  transition: all 0.2s ease;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+}
+
+.conceptChip:hover {
+  border-color: var(--color-stat-4);
+  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.15);
+  transform: translateY(-1px);
+}
+
+.conceptChip:active {
+  cursor: grabbing;
+}
+.emptyPool {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #a0aec0;
+  font-size: 0.875rem;
+  font-style: italic;
+  text-align: center;
+  padding: 1rem;
+  width: 100%;
+}
+.decisionsGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 1.5rem;
+}
+
+.decisionOption {
+  background: #f9fafb;
+  border: 2px solid #e5e7eb;
+  border-radius: 12px;
+  padding: 1.5rem;
+  text-align: left;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  position: relative;
+  overflow: hidden;
+  min-height: 200px;
+}
+
+.decisionOption:hover:not(:disabled) {
+  border-color: var(--color-primary);
+  background: white;
+  transform: translateY(-2px);
+  box-shadow: 0 4px 15px rgba(102, 126, 234, 0.15);
+}
+
+.decisionOption:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.optionHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1rem;
+}
+
+.optionNumber {
+  background: var(--color-stat-4);
+  color: white;
+  width: 2rem;
+  height: 2rem;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: 600;
+}
+
+.optionIcon {
+  color: var(--color-stat-4);
+  font-size: 1.2rem;
+}
+
+.optionContent {
+  margin-bottom: 1rem;
+  flex: 1;
+}
+
+.categoryName {
+  font-size: 1rem;
+  font-weight: 600;
+  color: var(--color-text);
+  margin: 0 0 1rem 0;
+}
+.categoryDropZone {
+  min-height: 100px;
+  border: 2px dashed #cbd5e0;
+  border-radius: 8px;
+  padding: 0.75rem;
+  transition: all 0.2s ease;
+  background: #f7fafc;
+}
+
+.categoryDropZone:hover {
+  border-color: var(--color-stat-4);
+  background: #f0f9ff;
+}
+
+.droppedConcepts {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.droppedConcept {
+  padding: 0.4rem 0.8rem;
+  background: #e0e7ff;
+  border: 1px solid #c7d2fe;
+  border-radius: 16px;
+  font-size: 0.8rem;
+  font-weight: 500;
+  color: var(--color-stat-4);
+  cursor: grab;
+  user-select: none;
+  transition: all 0.2s ease;
+}
+
+.droppedConcept:hover {
+  background: #c7d2fe;
+  transform: translateY(-1px);
+}
+
+.droppedConcept:active {
+  cursor: grabbing;
+}
+
+.emptyCategory {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #a0aec0;
+  font-size: 0.875rem;
+  font-style: italic;
+  text-align: center;
+  padding: 2rem 1rem;
+  height: 100%;
+}
+
+.optionFooter {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.5rem;
+}
+
+.optionStatus {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  color: #10b981;
+  font-size: 0.8rem;
+  font-weight: 500;
+}
+.statusIcon {
+  font-size: 0.7rem;
+}
+.completedActions {
+  display: flex;
+  justify-content: center;
+  padding-top: 1rem;
+}
+
+.tryAgainButton {
+  display: flex;
+  gap: 5px;
+  border: none;
+  box-shadow: 0 4px 15px rgba(102, 126, 234, 0.3);
+  transition: all 0.3s ease;
+}
+
+.tryAgainButton:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 6px 20px rgba(102, 126, 234, 0.4);
+}
+.resultSection {
+  margin-bottom: 2rem;
+}
+
+.resultCard {
+  background: white;
+  border-radius: 16px;
+  padding: 2rem;
+  border: 3px solid #e5e7eb;
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+  text-align: center;
+}
+
+.resultCard.approved {
+  border-color: #16a34a;
+  background: linear-gradient(135deg, #f0fdf4 0%, #dcfce7 100%);
+}
+
+.resultCard.rejected {
+  border-color: #dc2626;
+  background: linear-gradient(135deg, #fef2f2 0%, #fee2e2 100%);
+}
+
+.resultHeader {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 1.5rem;
+}
+
+.resultIcon {
+  font-size: 3rem;
+  color: #6b7280;
+}
+
+.resultCard.approved .resultIcon {
+  color: #16a34a;
+}
+
+.resultCard.rejected .resultIcon {
+  color: #dc2626;
+}
+
+.resultTitle {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: #374151;
+  margin: 0;
+}
+
+.resultCard.approved .resultTitle {
+  color: #166534;
+}
+
+.resultCard.rejected .resultTitle {
+  color: #991b1b;
+}
+
+.resultContent {
+  margin-bottom: 1.5rem;
+}
+
+.resultText {
+  font-size: 1.1rem;
+  line-height: 1.6;
+  color: #374151;
+  margin: 0;
+}
+
+.resultCard.approved .resultText {
+  color: #166534;
+}
+
+.resultCard.rejected .resultText {
+  color: #991b1b;
+}
+.verificationDetails {
+  margin-top: 1.5rem;
+  text-align: left;
+}
+
+.verificationSummary {
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
+  margin-bottom: 1rem;
+  padding: 1rem;
+  background: rgba(255, 255, 255, 0.5);
+  border-radius: 8px;
+}
+
+.correctCount {
+  color: #16a34a;
+  font-weight: 600;
+  font-size: 0.9rem;
+}
+
+.incorrectCount {
+  color: #dc2626;
+  font-weight: 600;
+  font-size: 0.9rem;
+}
+
+.incorrectDetails {
+  background: rgba(255, 255, 255, 0.7);
+  border-radius: 8px;
+  padding: 1rem;
+  border-left: 4px solid #dc2626;
+}
+
+.incorrectDetails h6 {
+  color: #991b1b;
+  font-size: 0.9rem;
+  font-weight: 600;
+  margin: 0 0 0.75rem 0;
+}
+
+.incorrectDetails ul {
+  margin: 0;
+  padding-left: 1.25rem;
+  list-style-type: none;
+}
+
+.incorrectDetails li {
+  color: #7f1d1d;
+  font-size: 0.85rem;
+  margin-bottom: 0.5rem;
+  position: relative;
+}
+
+.incorrectDetails li::before {
+  content: "→";
+  position: absolute;
+  left: -1rem;
+  color: #dc2626;
+  font-weight: bold;
+}
+
+.incorrectDetails strong {
+  color: #991b1b;
+}
+.resultFooter {
+  display: flex;
+  justify-content: center;
+}
+
+.successBadge,
+.failureBadge {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.75rem 1.5rem;
+  border-radius: 50px;
+  font-weight: 600;
+  font-size: 0.9rem;
+}
+
+.successBadge {
+  background: #16a34a;
+  color: white;
+}
+
+.failureBadge {
+  background: #dc2626;
+  color: white;
+}
+
+@media (max-width: 768px) {
+  .decisionsSectionHeader {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.5rem;
+  }
+
+  .resultCard {
+    padding: 1.5rem;
+  }
+
+  .resultIcon {
+    font-size: 2.5rem;
+  }
+
+  .resultTitle {
+    font-size: 1.25rem;
+  }
+}
