+++ src/shared/contexts/gamesContext/desafioClasificaionContext/DesafioClasificaionContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { DesafioClasificacionGameContextType } from "./DesafioClasificaionGameContext.type";
+
+export const DesafioGameContext = createContext<
+  DesafioClasificacionGameContextType | undefined
+>(undefined);
