+++ src/shared/contexts/gamesContext/desafioClasificaionContext/DesafioClasificaionGameContext.type.ts
@@ -1,0 +1,26 @@
+export interface DesafioClasificacionGameContextType {
+  // Estados del juego
+  gameStarted: boolean;
+  score: number;
+  gameStatus: "playing" | "won" | "lost";
+  draggedConcept: string | null;
+  conceptsInCategories: { [key: string]: string[] };
+  availableConcepts: string[];
+  verificationResults: {
+    correct: { concept: string; category: string }[];
+    incorrect: { concept: string; placedIn: string; shouldBe: string }[];
+    totalCorrect: number;
+    totalConcepts: number;
+  } | null;
+  totalCategories: number;
+  totalConcepts: number;
+
+  // Métodos
+  startGame: () => void;
+  resetGame: () => void;
+  verifyAnswers: () => void;
+  handleDragStart: (e: React.DragEvent, concept: string) => void;
+  handleDragOver: (e: React.DragEvent) => void;
+  handleDrop: (e: React.DragEvent, categoryId: string) => void;
+  handleDropToPool: (e: React.DragEvent) => void;
+}
